refactor: Rename `lib/plugins/aws/utils/resolveCfImportValue.js`
