+++ lib/plugins/aws/info/get-stack-info.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const BbPromise = require('bluebird');
-const resolveCfImportValue = require('../utils/resolveCfImportValue');
+const resolveCfImportValue = require('../utils/resolve-cf-import-value');
 const ServerlessError = require('../../../serverless-error');
 
 module.exports = {
