+++ lib/plugins/aws/invoke-local/index.js
@@ -21,7 +21,7 @@ const dirExists = require('../../../utils/fs/dir-exists');
 const fileExists = require('../../../utils/fs/file-exists');
 const isStandalone = require('../../../utils/is-standalone-executable');
 const ensureArtifact = require('../../../utils/ensure-artifact');
-const resolveCfImportValue = require('../utils/resolveCfImportValue');
+const resolveCfImportValue = require('../utils/resolve-cf-import-value');
 const resolveCfRefValue = require('../utils/resolveCfRefValue');
 const { writeText, progress, log } = require('@serverless/utils/log');
